+++ wp-content/plugins/newspack-newsletters/newspack-newsletters.php
@@ -4,10 +4,9 @@
  * Plugin URI:      https://newspack.pub
  * Description:     Newsletter authoring using the Gutenberg editor.
  * Author:          Automattic
- * Author URI:      https://newspack.pub
  * Text Domain:     newspack-newsletters
  * Domain Path:     /languages
- * Version:         1.11.2
+ * Version:         1.13.0
  *
  * @package         Newspack_Newsletters
  */