+++ wp-content/plugins/newspack-popups/includes/class-newspack-popups-inserter.php
@@ -49,7 +49,7 @@ final class Newspack_Popups_Inserter {
 
 		// Check if there's an overlay popup with matching category.
 		$category_overlay_popup = Newspack_Popups_Model::retrieve_category_overlay_popup();
-		if ( $category_overlay_popup ) {
+		if ( $category_overlay_popup && self::should_display( $category_overlay_popup ) ) {
 			array_push(
 				$popups_to_maybe_display,
 				$category_overlay_popup
@@ -133,8 +133,9 @@ final class Newspack_Popups_Inserter {
 	 * Process popups and insert into post and page content if needed.
 	 *
 	 * @param string $content The content of the post.
+	 * @param bool   $enqueue_assets Whether assets should be enqueued.
 	 */
-	public static function insert_popups_in_content( $content = '' ) {
+	public static function insert_popups_in_content( $content = '', $enqueue_assets = true ) {
 		// Avoid duplicate execution.
 		if ( true === self::$the_content_has_rendered ) {
 			return $content;
@@ -238,7 +239,9 @@ final class Newspack_Popups_Inserter {
 			$output = '<!-- wp:html -->' . Newspack_Popups_Model::generate_popup( $overlay_popup ) . '<!-- /wp:html -->' . $output;
 		}
 
-		self::enqueue_popup_assets();
+		if ( $enqueue_assets ) {
+			self::enqueue_popup_assets();
+		}
 		self::$the_content_has_rendered = true;
 		return $output;
 	}
@@ -305,70 +308,81 @@ final class Newspack_Popups_Inserter {
 	}
 
 	/**
+	 * Create the popup definition for sending to the API.
+	 *
+	 * @param object $popup A popup.
+	 */
+	public static function create_single_popup_access_payload( $popup ) {
+		$popup_id_string = Newspack_Popups_Model::canonize_popup_id( esc_attr( $popup['id'] ) );
+		$frequency       = $popup['options']['frequency'];
+		if ( 'inline' !== $popup['options']['placement'] && 'always' === $frequency ) {
+			$frequency = 'once';
+		}
+		return [
+			'id'  => $popup_id_string,
+			'f'   => $frequency,
+			'utm' => $popup['options']['utm_suppression'],
+			'n'   => \Newspack_Popups_Model::has_newsletter_prompt( $popup ),
+		];
+	}
+
+	/**
 	 * Add amp-access header code.
 	 *
-	 * @param object $popups The popup objects to handle.
+	 * The amp-access endpoint is also responsible for reporting visits, in order to minimise
+	 * the number of requests. For this reason it is placed on every page, not only those
+	 * with popups.
 	 */
-	public static function insert_popups_amp_access( $popups ) {
-		if ( is_admin() || self::assess_has_disabled_popups() ) {
+	public static function insert_popups_amp_access() {
+		if ( ! Newspack_Popups_Segmentation::is_tracking() ) {
 			return;
 		}
 
-		$popups                  = self::popups_for_post();
-		$endpoint                = str_replace(
-			'http://',
-			'//',
-			! Newspack_Popups::previewed_popup_id() && self::should_use_lightweight_api() ?
-				plugins_url( '../api/', __FILE__ ) :
-				get_rest_url( null, 'newspack-popups/v1/reader' )
-		);
-		$popups_access_providers = [];
+		$popups = self::popups_for_post();
+		// "Escape hatch" if there's a need to block adding amp-access for pages that have no campaigns.
+		if ( apply_filters( 'newspack_popups_suppress_insert_amp_access', false, $popups ) ) {
+			return;
+		}
 
-		$settings = \Newspack_Popups_Settings::get_settings();
+		$popups_access_provider = [
+			'namespace'     => 'popups',
+			'authorization' => esc_url( Newspack_Popups_Model::get_reader_endpoint() ) . '?cid=CLIENT_ID(' . Newspack_Popups_Segmentation::NEWSPACK_SEGMENTATION_CID_NAME . ')',
+			'noPingback'    => true,
+		];
 
+		$popups_configs = [];
 		foreach ( $popups as $popup ) {
-			// In test frequency cases (logged in site editor) and when previewing,
-			// fallback to authorization of true to avoid possible amp-access timeouts.
-			$authorization_fallback_response = (
-				( 'test' === $popup['options']['frequency'] || Newspack_Popups::previewed_popup_id() ) &&
-				is_user_logged_in() &&
-				current_user_can( 'edit_others_pages' )
-			);
+			$popups_configs[] = self::create_single_popup_access_payload( $popup );
+		}
 
-			$authorization =
-				esc_url( $endpoint ) .
-				'?popup_id=' . esc_attr( $popup['id'] ) .
-				'&rid=READER_ID&url=CANONICAL_URL&RANDOM&' .
-				'frequency=' . esc_attr( $popup['options']['frequency'] ) .
-				'&placement=' . esc_attr( $popup['options']['placement'] ) .
-				'&utm_suppression=' . esc_attr( $popup['options']['utm_suppression'] ) .
-				'&suppress_newsletter_campaigns=' . esc_attr( $settings['suppress_newsletter_campaigns'] ) .
-				'&suppress_all_newsletter_campaigns_if_one_dismissed=' . esc_attr( $settings['suppress_all_newsletter_campaigns_if_one_dismissed'] ) .
-				'&has_newsletter_prompt=' . esc_attr( \Newspack_Popups_Model::has_newsletter_prompt( $popup ) );
-
-			$amp_access_provider = array(
-				'namespace'                     => 'popup_' . $popup['id'],
-				'authorization'                 => $authorization,
-				'noPingback'                    => true,
-				'authorizationTimeout'          => 10000, // For development purposes. If #development=1 is appended to URL the maximum timeout for amp-access is raised to 10s.
-				'authorizationFallbackResponse' => array(
-					'displayPopup' => $authorization_fallback_response,
-				),
-			);
-
-			array_push(
-				$popups_access_providers,
-				$amp_access_provider
+		$categories   = get_the_category();
+		$category_ids = '';
+		if ( ! empty( $categories ) ) {
+			$category_ids = implode(
+				',',
+				array_map(
+					function( $cat ) {
+						return $cat->term_id;
+					},
+					$categories
+				)
 			);
 		}
 
-		if ( ! empty( $popups_access_providers ) ) {
-			?>
-			<script id="amp-access" type="application/json">
-				<?php echo wp_json_encode( $popups_access_providers ); ?>
-			</script>
-			<?php
-		}
+		$popups_access_provider['authorization'] .= '&popups=' . wp_json_encode( $popups_configs );
+		$popups_access_provider['authorization'] .= '&settings=' . wp_json_encode( \Newspack_Popups_Settings::get_settings() );
+		$popups_access_provider['authorization'] .= '&visit=' . wp_json_encode(
+			[
+				'post_id'    => esc_attr( get_the_ID() ),
+				'categories' => esc_attr( $category_ids ),
+				'is_post'    => is_single(),
+			]
+		);
+		?>
+		<script id="amp-access" type="application/json">
+			<?php echo wp_json_encode( $popups_access_provider ); ?>
+		</script>
+		<?php
 	}
 
 	/**
@@ -384,7 +398,7 @@ final class Newspack_Popups_Inserter {
 	 * Register and enqueue all required AMP scripts, if needed.
 	 */
 	public static function register_amp_scripts() {
-		if ( self::assess_has_disabled_popups() ) {
+		if ( ! Newspack_Popups_Segmentation::is_tracking() ) {
 			return;
 		}
 		if ( ! is_admin() && ! wp_script_is( 'amp-runtime', 'registered' ) ) {
@@ -423,9 +437,7 @@ final class Newspack_Popups_Inserter {
 	public static function assess_is_post( $popup ) {
 		if (
 			// Inline Pop-ups can only appear in Posts.
-			'inline' === $popup['options']['placement'] ||
-			// Pop-ups triggered by scroll position can only appear on Posts.
-			'scroll' === $popup['options']['trigger_type']
+			'inline' === $popup['options']['placement']
 		) {
 			return is_single();
 		}
@@ -477,15 +489,6 @@ final class Newspack_Popups_Inserter {
 		return self::assess_is_post( $popup ) &&
 			self::assess_test_mode( $popup ) &&
 			self::assess_categories_filter( $popup );
-	}
-
-	/**
-	 * Should the lightweight API be used.
-	 *
-	 * @return bool Should lightweight API be used.
-	 */
-	public static function should_use_lightweight_api() {
-		return file_exists( WP_CONTENT_DIR . '/../newspack-popups-config.php' );
 	}
 }
 $newspack_popups_inserter = new Newspack_Popups_Inserter();